feat: add coverage support
