+++ Justfile
@@ -20,3 +20,11 @@ lint:
 # Test all
 test:
     cargo test --all
+
+# Test with coverage
+test-cov:
+    rm -rf .cov
+    rm -rf htmlcov
+    RUSTFLAGS="-Cinstrument-coverage" LLVM_PROFILE_FILE=".cov/test-%p-%m.profraw" cargo build
+    RUSTFLAGS="-Cinstrument-coverage" LLVM_PROFILE_FILE=".cov/test-%p-%m.profraw" cargo test
+    grcov . --binary-path ./target/debug/ -s . -t html --branch --ignore-not-existing --ignore "/*" -o htmlcov
