+++ crates/gdtoolkit-gdscript-lexer/src/lib.rs
@@ -5,8 +5,10 @@
 mod debug;
 mod error;
 mod lexer;
-mod read;
+mod read_context;
+mod readers;
 mod serializer;
+mod span;
 mod token;
 
 #[cfg(test)]
@@ -15,7 +17,7 @@ mod tests;
 pub use debug::CompactTokenView;
 pub use error::{Error, Result};
 pub use lexer::{GdScriptLexer, GdScriptLexerOutput};
-pub use read::TokenReaderContext;
+pub use read_context::TokenReaderContext;
 pub use serializer::{GdScriptLexerOutputFormat, GdScriptLexerOutputSerializer};
 pub use token::{
     FloatType, IndentationType, IntType, Keyword, NewLine, Operator, Punct, QuoteMode, Token, Value,
